refactor: move e2ei structs into a unified module tree
