+++ crypto-ffi/src/generic/mod.rs
@@ -1,250 +1,7 @@
-use self::context::CoreCryptoContext;
 use crate::{Ciphersuite, Ciphersuites, ClientId, CoreCrypto, CoreCryptoError, CoreCryptoResult, CredentialType};
 
 pub mod context;
 mod epoch_observer;
-
-#[derive(Debug, uniffi::Object)]
-/// See [core_crypto::e2e_identity::E2eiEnrollment]
-pub struct E2eiEnrollment(std::sync::Arc<async_lock::RwLock<core_crypto::prelude::E2eiEnrollment>>);
-
-#[uniffi::export]
-impl E2eiEnrollment {
-    /// See [core_crypto::e2e_identity::E2eiEnrollment::directory_response]
-    pub async fn directory_response(&self, directory: Vec<u8>) -> CoreCryptoResult<AcmeDirectory> {
-        Ok(self
-            .0
-            .write()
-            .await
-            .directory_response(directory)
-            .map(AcmeDirectory::from)?)
-    }
-
-    /// See [core_crypto::e2e_identity::E2eiEnrollment::new_account_request]
-    pub async fn new_account_request(&self, previous_nonce: String) -> CoreCryptoResult<Vec<u8>> {
-        Ok(self.0.read().await.new_account_request(previous_nonce)?)
-    }
-
-    /// See [core_crypto::e2e_identity::E2eiEnrollment::new_account_response]
-    pub async fn new_account_response(&self, account: Vec<u8>) -> CoreCryptoResult<()> {
-        Ok(self.0.write().await.new_account_response(account)?)
-    }
-
-    /// See [core_crypto::e2e_identity::E2eiEnrollment::new_order_request]
-    #[allow(clippy::too_many_arguments)]
-    pub async fn new_order_request(&self, previous_nonce: String) -> CoreCryptoResult<Vec<u8>> {
-        Ok(self.0.read().await.new_order_request(previous_nonce)?)
-    }
-
-    /// See [core_crypto::e2e_identity::E2eiEnrollment::new_order_response]
-    pub async fn new_order_response(&self, order: Vec<u8>) -> CoreCryptoResult<NewAcmeOrder> {
-        Ok(self.0.read().await.new_order_response(order)?.into())
-    }
-
-    /// See [core_crypto::e2e_identity::E2eiEnrollment::new_authz_request]
-    pub async fn new_authz_request(&self, url: String, previous_nonce: String) -> CoreCryptoResult<Vec<u8>> {
-        Ok(self.0.read().await.new_authz_request(url, previous_nonce)?)
-    }
-
-    /// See [core_crypto::e2e_identity::E2eiEnrollment::new_authz_response]
-    pub async fn new_authz_response(&self, authz: Vec<u8>) -> CoreCryptoResult<NewAcmeAuthz> {
-        Ok(self.0.write().await.new_authz_response(authz)?.into())
-    }
-
-    #[allow(clippy::too_many_arguments)]
-    /// See [core_crypto::e2e_identity::E2eiEnrollment::create_dpop_token]
-    pub async fn create_dpop_token(&self, expiry_secs: u32, backend_nonce: String) -> CoreCryptoResult<String> {
-        Ok(self.0.read().await.create_dpop_token(expiry_secs, backend_nonce)?)
-    }
-
-    /// See [core_crypto::e2e_identity::E2eiEnrollment::new_dpop_challenge_request]
-    pub async fn new_dpop_challenge_request(
-        &self,
-        access_token: String,
-        previous_nonce: String,
-    ) -> CoreCryptoResult<Vec<u8>> {
-        Ok(self
-            .0
-            .read()
-            .await
-            .new_dpop_challenge_request(access_token, previous_nonce)?)
-    }
-
-    /// See [core_crypto::e2e_identity::E2eiEnrollment::new_dpop_challenge_response]
-    pub async fn new_dpop_challenge_response(&self, challenge: Vec<u8>) -> CoreCryptoResult<()> {
-        Ok(self.0.read().await.new_dpop_challenge_response(challenge)?)
-    }
-
-    /// See [core_crypto::e2e_identity::E2eiEnrollment::new_oidc_challenge_request]
-    pub async fn new_oidc_challenge_request(
-        &self,
-        id_token: String,
-        refresh_token: String,
-        previous_nonce: String,
-    ) -> CoreCryptoResult<Vec<u8>> {
-        Ok(self
-            .0
-            .write()
-            .await
-            .new_oidc_challenge_request(id_token, refresh_token, previous_nonce)?)
-    }
-
-    /// See [core_crypto::e2e_identity::E2eiEnrollment::new_oidc_challenge_response]
-    pub async fn context_new_oidc_challenge_response(
-        &self,
-        cc: std::sync::Arc<CoreCryptoContext>,
-        challenge: Vec<u8>,
-    ) -> CoreCryptoResult<()> {
-        self.0
-            .write()
-            .await
-            .new_oidc_challenge_response(&cc.context.mls_provider().await?, challenge)
-            .await?;
-        Ok(())
-    }
-
-    /// See [core_crypto::e2e_identity::E2eiEnrollment::check_order_request]
-    pub async fn check_order_request(&self, order_url: String, previous_nonce: String) -> CoreCryptoResult<Vec<u8>> {
-        Ok(self.0.read().await.check_order_request(order_url, previous_nonce)?)
-    }
-
-    /// See [core_crypto::e2e_identity::E2eiEnrollment::check_order_response]
-    pub async fn check_order_response(&self, order: Vec<u8>) -> CoreCryptoResult<String> {
-        Ok(self.0.write().await.check_order_response(order)?)
-    }
-
-    /// See [core_crypto::prelude::E2eiEnrollment::finalize_request]
-    pub async fn finalize_request(&self, previous_nonce: String) -> CoreCryptoResult<Vec<u8>> {
-        Ok(self.0.write().await.finalize_request(previous_nonce)?)
-    }
-
-    /// See [core_crypto::prelude::E2eiEnrollment::finalize_response]
-    pub async fn finalize_response(&self, finalize: Vec<u8>) -> CoreCryptoResult<String> {
-        Ok(self.0.write().await.finalize_response(finalize)?)
-    }
-
-    /// See [core_crypto::prelude::E2eiEnrollment::certificate_request]
-    pub async fn certificate_request(&self, previous_nonce: String) -> CoreCryptoResult<Vec<u8>> {
-        Ok(self.0.write().await.certificate_request(previous_nonce)?)
-    }
-
-    /// See [core_crypto::prelude::E2eiEnrollment::get_refresh_token]
-    pub async fn get_refresh_token(&self) -> CoreCryptoResult<String> {
-        Ok(self.0.read().await.get_refresh_token().map(Into::into)?)
-    }
-}
-
-#[derive(Debug, uniffi::Record)]
-/// See [core_crypto::e2e_identity::types::E2eiAcmeDirectory]
-pub struct AcmeDirectory {
-    pub new_nonce: String,
-    pub new_account: String,
-    pub new_order: String,
-    pub revoke_cert: String,
-}
-
-impl From<core_crypto::prelude::E2eiAcmeDirectory> for AcmeDirectory {
-    fn from(directory: core_crypto::prelude::E2eiAcmeDirectory) -> Self {
-        Self {
-            new_nonce: directory.new_nonce,
-            new_account: directory.new_account,
-            new_order: directory.new_order,
-            revoke_cert: directory.revoke_cert,
-        }
-    }
-}
-
-impl From<AcmeDirectory> for core_crypto::prelude::E2eiAcmeDirectory {
-    fn from(directory: AcmeDirectory) -> Self {
-        Self {
-            new_nonce: directory.new_nonce,
-            new_account: directory.new_account,
-            new_order: directory.new_order,
-            revoke_cert: directory.revoke_cert,
-        }
-    }
-}
-
-#[derive(Debug, uniffi::Record)]
-/// See [core_crypto::e2e_identity::types::E2eiNewAcmeOrder]
-pub struct NewAcmeOrder {
-    pub delegate: Vec<u8>,
-    pub authorizations: Vec<String>,
-}
-
-impl From<core_crypto::prelude::E2eiNewAcmeOrder> for NewAcmeOrder {
-    fn from(new_order: core_crypto::prelude::E2eiNewAcmeOrder) -> Self {
-        Self {
-            delegate: new_order.delegate,
-            authorizations: new_order.authorizations,
-        }
-    }
-}
-
-impl From<NewAcmeOrder> for core_crypto::prelude::E2eiNewAcmeOrder {
-    fn from(new_order: NewAcmeOrder) -> Self {
-        Self {
-            delegate: new_order.delegate,
-            authorizations: new_order.authorizations,
-        }
-    }
-}
-
-#[derive(Debug, uniffi::Record)]
-/// See [core_crypto::e2e_identity::types::E2eiNewAcmeAuthz]
-pub struct NewAcmeAuthz {
-    pub identifier: String,
-    pub keyauth: Option<String>,
-    pub challenge: AcmeChallenge,
-}
-
-impl From<core_crypto::prelude::E2eiNewAcmeAuthz> for NewAcmeAuthz {
-    fn from(new_authz: core_crypto::prelude::E2eiNewAcmeAuthz) -> Self {
-        Self {
-            identifier: new_authz.identifier,
-            keyauth: new_authz.keyauth,
-            challenge: new_authz.challenge.into(),
-        }
-    }
-}
-
-impl From<NewAcmeAuthz> for core_crypto::prelude::E2eiNewAcmeAuthz {
-    fn from(new_authz: NewAcmeAuthz) -> Self {
-        Self {
-            identifier: new_authz.identifier,
-            keyauth: new_authz.keyauth,
-            challenge: new_authz.challenge.into(),
-        }
-    }
-}
-
-#[derive(Debug, uniffi::Record)]
-/// See [core_crypto::e2e_identity::types::E2eiAcmeChallenge]
-pub struct AcmeChallenge {
-    pub delegate: Vec<u8>,
-    pub url: String,
-    pub target: String,
-}
-
-impl From<core_crypto::prelude::E2eiAcmeChallenge> for AcmeChallenge {
-    fn from(chall: core_crypto::prelude::E2eiAcmeChallenge) -> Self {
-        Self {
-            delegate: chall.delegate,
-            url: chall.url,
-            target: chall.target,
-        }
-    }
-}
-
-impl From<AcmeChallenge> for core_crypto::prelude::E2eiAcmeChallenge {
-    fn from(chall: AcmeChallenge) -> Self {
-        Self {
-            delegate: chall.delegate,
-            url: chall.url,
-            target: chall.target,
-        }
-    }
-}
 
 #[cfg(test)]
 mod tests {
