+++ crypto-ffi/src/lib.rs
@@ -38,20 +38,26 @@ pub(crate) use ciphersuite::lower_ciphersuites;
 pub use ciphersuite::{Ciphersuite, Ciphersuites};
 pub use client_id::{ClientId, FfiClientId};
 pub use configuration::{ConversationConfiguration, CustomConfiguration, WirePolicy};
-pub use core_crypto::CoreCrypto;
 pub(crate) use core_crypto::conversation::ConversationId;
-pub use core_crypto::e2ei::E2eiDumpedPkiEnv;
-pub use core_crypto::logger::{CoreCryptoLogLevel, CoreCryptoLogger};
+pub use core_crypto::{
+    CoreCrypto,
+    e2ei::E2eiDumpedPkiEnv,
+    logger::{CoreCryptoLogLevel, CoreCryptoLogger},
+    mls_transport::{MlsTransport, MlsTransportResponse},
+};
 #[cfg(not(target_family = "wasm"))]
-pub use core_crypto::logger::{set_logger, set_logger_only, set_max_log_level};
-pub use core_crypto::mls_transport::{MlsTransport, MlsTransportResponse};
-#[cfg(not(target_family = "wasm"))]
-pub use core_crypto::{core_crypto_deferred_init, core_crypto_new};
+pub use core_crypto::{
+    core_crypto_deferred_init, core_crypto_new,
+    logger::{set_logger, set_logger_only, set_max_log_level},
+};
 pub use credential_type::CredentialType;
 pub use crl::{CrlRegistration, NewCrlDistributionPoints};
 pub use database_key::{DatabaseKey, migrate_db_key_type_to_bytes};
 pub use decrypted_message::{BufferedDecryptedMessage, DecryptedMessage};
-pub use e2ei::E2eiConversationState;
+pub use e2ei::{
+    E2eiConversationState, acme_challenge::AcmeChallenge, acme_directory::AcmeDirectory, enrollment::E2eiEnrollment,
+    new_acme_authz::NewAcmeAuthz, new_acme_order::NewAcmeOrder,
+};
 #[cfg(feature = "proteus")]
 pub use error::proteus::ProteusError;
 pub use error::{CoreCryptoResult, core_crypto::CoreCryptoError, mls::MlsError};
