+++ crypto-ffi/src/wasm/context/mod.rs
@@ -1,5 +1,5 @@
 use crate::InternalError;
-use crate::wasm::lower_ciphersuites;
+use crate::lower_ciphersuites;
 use crate::{
     Ciphersuite, ConversationConfiguration, CoreCrypto, CoreCryptoError, CoreCryptoResult, CredentialType,
     CustomConfiguration, DecryptedMessage, FfiClientId, WasmCryptoResult, WelcomeBundle,
