+++ crypto-ffi/src/wasm/mod.rs
@@ -2,426 +2,3 @@
 pub mod context;
 mod epoch_observer;
 mod utils;
-
-use std::sync::Arc;
-
-use futures_util::future::TryFutureExt;
-use js_sys::{Promise, Uint8Array};
-use utils::*;
-use wasm_bindgen::prelude::*;
-use wasm_bindgen_futures::future_to_promise;
-
-use crate::{CoreCrypto, CoreCryptoError, InternalError, WasmCryptoResult, lower_ciphersuites};
-
-#[derive(Debug)]
-#[wasm_bindgen(js_name = FfiWireE2EIdentity)]
-#[repr(transparent)]
-pub struct E2eiEnrollment(pub(super) Arc<async_lock::RwLock<core_crypto::prelude::E2eiEnrollment>>);
-
-#[wasm_bindgen(js_class = FfiWireE2EIdentity)]
-impl E2eiEnrollment {
-    /// See [core_crypto::e2e_identity::WireE2eIdentity::directory_response]
-    pub fn directory_response(&mut self, directory: Vec<u8>) -> Promise {
-        let this = self.0.clone();
-        future_to_promise(
-            async move {
-                let mut this = this.write().await;
-                let directory: AcmeDirectory = this.directory_response(directory)?.into();
-                WasmCryptoResult::Ok(directory.into())
-            }
-            .err_into(),
-        )
-    }
-
-    /// See [core_crypto::e2e_identity::WireE2eIdentity::new_account_request]
-    pub fn new_account_request(&self, previous_nonce: String) -> Promise {
-        let this = self.0.clone();
-        future_to_promise(
-            async move {
-                let this = this.read().await;
-                let new_account: Vec<u8> = this.new_account_request(previous_nonce)?;
-                WasmCryptoResult::Ok(Uint8Array::from(new_account.as_slice()).into())
-            }
-            .err_into(),
-        )
-    }
-
-    /// See [core_crypto::e2e_identity::WireE2eIdentity::new_account_response]
-    pub fn new_account_response(&mut self, account: Uint8Array) -> Promise {
-        let this = self.0.clone();
-        future_to_promise(
-            async move {
-                let mut this = this.write().await;
-                this.new_account_response(account.to_vec())?;
-                WasmCryptoResult::Ok(JsValue::UNDEFINED)
-            }
-            .err_into(),
-        )
-    }
-
-    /// See [core_crypto::e2e_identity::WireE2eIdentity::new_order_request]
-    pub fn new_order_request(&self, previous_nonce: String) -> Promise {
-        let this = self.0.clone();
-        future_to_promise(
-            async move {
-                let this = this.read().await;
-                let new_order = this.new_order_request(previous_nonce)?;
-                WasmCryptoResult::Ok(Uint8Array::from(new_order.as_slice()).into())
-            }
-            .err_into(),
-        )
-    }
-
-    /// See [core_crypto::e2e_identity::WireE2eIdentity::new_order_response]
-    pub fn new_order_response(&self, order: Uint8Array) -> Promise {
-        let this = self.0.clone();
-        future_to_promise(
-            async move {
-                let this = this.read().await;
-                let order: NewAcmeOrder = this.new_order_response(order.to_vec())?.into();
-                WasmCryptoResult::Ok(order.into())
-            }
-            .err_into(),
-        )
-    }
-
-    /// See [core_crypto::e2e_identity::WireE2eIdentity::new_authz_request]
-    pub fn new_authz_request(&self, url: String, previous_nonce: String) -> Promise {
-        let this = self.0.clone();
-        future_to_promise(
-            async move {
-                let this = this.read().await;
-                let new_authz = this.new_authz_request(url, previous_nonce)?;
-                WasmCryptoResult::Ok(Uint8Array::from(new_authz.as_slice()).into())
-            }
-            .err_into(),
-        )
-    }
-
-    /// See [core_crypto::e2e_identity::WireE2eIdentity::new_authz_response]
-    pub fn new_authz_response(&mut self, authz: Uint8Array) -> Promise {
-        let this = self.0.clone();
-        future_to_promise(
-            async move {
-                let mut this = this.write().await;
-                let authz: NewAcmeAuthz = this.new_authz_response(authz.to_vec())?.into();
-                WasmCryptoResult::Ok(authz.into())
-            }
-            .err_into(),
-        )
-    }
-
-    /// See [core_crypto::e2e_identity::WireE2eIdentity::create_dpop_token]
-    pub fn create_dpop_token(&self, expiry_secs: u32, backend_nonce: String) -> Promise {
-        let this = self.0.clone();
-        future_to_promise(
-            async move {
-                let this = this.read().await;
-                let dpop_token = this.create_dpop_token(expiry_secs, backend_nonce)?;
-                WasmCryptoResult::Ok(Uint8Array::from(dpop_token.as_bytes()).into())
-            }
-            .err_into(),
-        )
-    }
-
-    /// See [core_crypto::e2e_identity::WireE2eIdentity::new_dpop_challenge_request]
-    pub fn new_dpop_challenge_request(&self, access_token: String, previous_nonce: String) -> Promise {
-        let this = self.0.clone();
-        future_to_promise(
-            async move {
-                let this = this.read().await;
-                let chall = this.new_dpop_challenge_request(access_token, previous_nonce)?;
-                WasmCryptoResult::Ok(Uint8Array::from(chall.as_slice()).into())
-            }
-            .err_into(),
-        )
-    }
-
-    /// See [core_crypto::e2e_identity::WireE2eIdentity::new_dpop_challenge_response]
-    pub fn new_dpop_challenge_response(&self, challenge: Uint8Array) -> Promise {
-        let this = self.0.clone();
-        future_to_promise(
-            async move {
-                let this = this.read().await;
-                this.new_dpop_challenge_response(challenge.to_vec())?;
-                WasmCryptoResult::Ok(JsValue::UNDEFINED)
-            }
-            .err_into(),
-        )
-    }
-
-    /// See [core_crypto::e2e_identity::WireE2eIdentity::new_oidc_challenge_request]
-    pub fn new_oidc_challenge_request(&mut self, id_token: String, previous_nonce: String) -> Promise {
-        let this = self.0.clone();
-        future_to_promise(
-            async move {
-                let mut this = this.write().await;
-                let chall = this.new_oidc_challenge_request(id_token, previous_nonce)?;
-                WasmCryptoResult::Ok(Uint8Array::from(chall.as_slice()).into())
-            }
-            .err_into(),
-        )
-    }
-
-    /// See [core_crypto::e2e_identity::WireE2eIdentity::new_oidc_challenge_response]
-    pub fn new_oidc_challenge_response(&mut self, challenge: Uint8Array) -> Promise {
-        let this = self.0.clone();
-        future_to_promise(
-            async move {
-                let mut this = this.write().await;
-                this.new_oidc_challenge_response(challenge.to_vec()).await?;
-                WasmCryptoResult::Ok(JsValue::UNDEFINED)
-            }
-            .err_into(),
-        )
-    }
-
-    /// See [core_crypto::e2e_identity::WireE2eIdentity::check_order_request]
-    pub fn check_order_request(&self, order_url: String, previous_nonce: String) -> Promise {
-        let this = self.0.clone();
-        future_to_promise(
-            async move {
-                let this = this.read().await;
-                let new_order = this.check_order_request(order_url, previous_nonce)?;
-                WasmCryptoResult::Ok(Uint8Array::from(new_order.as_slice()).into())
-            }
-            .err_into(),
-        )
-    }
-
-    /// See [core_crypto::e2e_identity::WireE2eIdentity::check_order_response]
-    pub fn check_order_response(&mut self, order: Uint8Array) -> Promise {
-        let this = self.0.clone();
-        future_to_promise(
-            async move {
-                let mut this = this.write().await;
-                WasmCryptoResult::Ok(this.check_order_response(order.to_vec())?.into())
-            }
-            .err_into(),
-        )
-    }
-
-    /// See [core_crypto::e2e_identity::WireE2eIdentity::finalize_request]
-    pub fn finalize_request(&mut self, previous_nonce: String) -> Promise {
-        let this = self.0.clone();
-        future_to_promise(
-            async move {
-                let mut this = this.write().await;
-                let finalize = this.finalize_request(previous_nonce)?;
-                WasmCryptoResult::Ok(Uint8Array::from(finalize.as_slice()).into())
-            }
-            .err_into(),
-        )
-    }
-
-    /// See [core_crypto::e2e_identity::WireE2eIdentity::finalize_response]
-    pub fn finalize_response(&mut self, finalize: Uint8Array) -> Promise {
-        let this = self.0.clone();
-        future_to_promise(
-            async move {
-                let mut this = this.write().await;
-                WasmCryptoResult::Ok(this.finalize_response(finalize.to_vec())?.into())
-            }
-            .err_into(),
-        )
-    }
-
-    /// See [core_crypto::e2e_identity::WireE2eIdentity::certificate_request]
-    pub fn certificate_request(&mut self, previous_nonce: String) -> Promise {
-        let this = self.0.clone();
-        future_to_promise(
-            async move {
-                let mut this = this.write().await;
-                let certificate_req = this.certificate_request(previous_nonce)?;
-                WasmCryptoResult::Ok(Uint8Array::from(certificate_req.as_slice()).into())
-            }
-            .err_into(),
-        )
-    }
-}
-
-/// Holds URLs of all the standard ACME endpoint supported on an ACME server.
-/// @see https://www.rfc-editor.org/rfc/rfc8555.html#section-7.1.1
-#[wasm_bindgen(skip_jsdoc, getter_with_clone)]
-#[derive(Debug, Clone, serde::Serialize, serde::Deserialize)]
-pub struct AcmeDirectory {
-    /// URL for fetching a new nonce. Use this only for creating a new account.
-    #[wasm_bindgen(readonly)]
-    pub new_nonce: String,
-    /// URL for creating a new account.
-    #[wasm_bindgen(readonly)]
-    pub new_account: String,
-    /// URL for creating a new order.
-    #[wasm_bindgen(readonly)]
-    pub new_order: String,
-    /// Revocation URL
-    #[wasm_bindgen(readonly)]
-    pub revoke_cert: String,
-}
-
-impl From<core_crypto::prelude::E2eiAcmeDirectory> for AcmeDirectory {
-    fn from(directory: core_crypto::prelude::E2eiAcmeDirectory) -> Self {
-        Self {
-            new_nonce: directory.new_nonce,
-            new_account: directory.new_account,
-            new_order: directory.new_order,
-            revoke_cert: directory.revoke_cert,
-        }
-    }
-}
-
-impl From<AcmeDirectory> for core_crypto::prelude::E2eiAcmeDirectory {
-    fn from(directory: AcmeDirectory) -> Self {
-        Self {
-            new_nonce: directory.new_nonce,
-            new_account: directory.new_account,
-            new_order: directory.new_order,
-            revoke_cert: directory.revoke_cert,
-        }
-    }
-}
-
-/// Result of an order creation.
-/// @see https://www.rfc-editor.org/rfc/rfc8555.html#section-7.4
-#[wasm_bindgen(skip_jsdoc)]
-#[derive(Debug, Clone, serde::Serialize, serde::Deserialize)]
-pub struct NewAcmeOrder {
-    /// Contains raw JSON data of this order. This is parsed by the underlying Rust library hence should not be accessed
-    #[wasm_bindgen(readonly, getter_with_clone)]
-    pub delegate: Vec<u8>,
-    authorizations: ArrayOfByteArray,
-}
-
-#[wasm_bindgen]
-impl NewAcmeOrder {
-    #[wasm_bindgen(getter)]
-    pub fn authorizations(&self) -> Vec<Uint8Array> {
-        self.authorizations.clone().into()
-    }
-}
-
-impl From<core_crypto::prelude::E2eiNewAcmeOrder> for NewAcmeOrder {
-    fn from(new_order: core_crypto::prelude::E2eiNewAcmeOrder) -> Self {
-        Self {
-            delegate: new_order.delegate,
-            authorizations: new_order
-                .authorizations
-                .into_iter()
-                .map(String::into_bytes)
-                .collect::<Vec<_>>()
-                .into(),
-        }
-    }
-}
-
-impl TryFrom<NewAcmeOrder> for core_crypto::prelude::E2eiNewAcmeOrder {
-    type Error = CoreCryptoError;
-
-    fn try_from(new_order: NewAcmeOrder) -> WasmCryptoResult<Self> {
-        let authorizations = new_order
-            .authorizations
-            .0
-            .into_iter()
-            .map(String::from_utf8)
-            .collect::<Result<Vec<String>, _>>()
-            .map_err(|_| InternalError::Other("invalid authorization string: not utf8".into()))?;
-        Ok(Self {
-            delegate: new_order.delegate,
-            authorizations,
-        })
-    }
-}
-
-/// Result of an authorization creation.
-/// @see https://www.rfc-editor.org/rfc/rfc8555.html#section-7.5
-#[wasm_bindgen(skip_jsdoc, getter_with_clone)]
-#[derive(Debug, Clone, serde::Serialize, serde::Deserialize)]
-pub struct NewAcmeAuthz {
-    /// DNS entry associated with those challenge
-    #[wasm_bindgen(readonly)]
-    pub identifier: String,
-    /// ACME challenge + ACME key thumbprint
-    #[wasm_bindgen(readonly)]
-    pub keyauth: Option<String>,
-    /// Associated ACME Challenge
-    #[wasm_bindgen(readonly)]
-    pub challenge: AcmeChallenge,
-}
-
-impl From<core_crypto::prelude::E2eiNewAcmeAuthz> for NewAcmeAuthz {
-    fn from(authz: core_crypto::prelude::E2eiNewAcmeAuthz) -> Self {
-        Self {
-            identifier: authz.identifier,
-            keyauth: authz.keyauth,
-            challenge: authz.challenge.into(),
-        }
-    }
-}
-
-impl From<NewAcmeAuthz> for core_crypto::prelude::E2eiNewAcmeAuthz {
-    fn from(authz: NewAcmeAuthz) -> Self {
-        Self {
-            identifier: authz.identifier,
-            keyauth: authz.keyauth,
-            challenge: authz.challenge.into(),
-        }
-    }
-}
-
-/// For creating a challenge.
-/// @see https://www.rfc-editor.org/rfc/rfc8555.html#section-7.5.1
-#[wasm_bindgen(skip_jsdoc, getter_with_clone)]
-#[derive(Debug, Clone, serde::Serialize, serde::Deserialize)]
-pub struct AcmeChallenge {
-    /// Contains raw JSON data of this challenge. This is parsed by the underlying Rust library hence should not be accessed
-    #[wasm_bindgen(readonly)]
-    pub delegate: Vec<u8>,
-    /// URL of this challenge
-    #[wasm_bindgen(readonly)]
-    pub url: String,
-    /// Non-standard, Wire specific claim. Indicates the consumer from where it should get the challenge proof.
-    /// Either from wire-server "/access-token" endpoint in case of a DPoP challenge, or from an OAuth token endpoint for an OIDC challenge
-    #[wasm_bindgen(readonly)]
-    pub target: String,
-}
-
-impl From<core_crypto::prelude::E2eiAcmeChallenge> for AcmeChallenge {
-    fn from(chall: core_crypto::prelude::E2eiAcmeChallenge) -> Self {
-        Self {
-            delegate: chall.delegate,
-            url: chall.url,
-            target: chall.target,
-        }
-    }
-}
-
-impl From<AcmeChallenge> for core_crypto::prelude::E2eiAcmeChallenge {
-    fn from(chall: AcmeChallenge) -> Self {
-        Self {
-            delegate: chall.delegate,
-            url: chall.url,
-            target: chall.target,
-        }
-    }
-}
-
-#[derive(Debug, Clone, Copy, PartialEq, Eq, serde::Serialize, serde::Deserialize)]
-#[repr(u8)]
-/// see [core_crypto::prelude::E2eiConversationState]
-enum E2eiConversationState {
-    Verified = 1,
-    /// Some clients are either still Basic or their certificate is expired
-    NotVerified = 2,
-    /// All clients are still Basic. If all client have expired certificates, [E2eiConversationState::NotVerified] is returned.
-    NotEnabled = 3,
-}
-
-impl From<core_crypto::prelude::E2eiConversationState> for E2eiConversationState {
-    fn from(state: core_crypto::prelude::E2eiConversationState) -> Self {
-        match state {
-            core_crypto::prelude::E2eiConversationState::Verified => Self::Verified,
-            core_crypto::prelude::E2eiConversationState::NotVerified => Self::NotVerified,
-            core_crypto::prelude::E2eiConversationState::NotEnabled => Self::NotEnabled,
-        }
-    }
-}
